moved building card header to separate component
