+++ src/components/all-buildings/BuildingCard.tsx
@@ -1,13 +1,10 @@
-import { useContext, MutableRefObject } from "react";
-import { deleteBuilding, saveBuilding } from "../../utils/firestoreUtils";
+import { MutableRefObject } from "react";
 import { calculateDaysAgo, willShowAvailTable } from "../../utils/generalUtils";
 import { useAuth } from "../../contexts/AuthContext";
-import { ModalContext, ModalState } from "../../contexts/ModalContext";
 import { AddressAndPhone } from "../shared/AddressAndPhone";
+import BuildingCardHeader from "./BuildingCardHeader";
 import BuildingCardNote from "./BuildingCardNote";
 import BuildingDataTable from "../shared/BuildingDataTable";
-import SaveButton from "../shared/SaveButton";
-import WebsiteButton from "../shared/WebsiteButton";
 import {
   BuildingCardEnum,
   listingStatusEnum,
@@ -17,12 +14,10 @@ import {
 import IBuilding from "../../interfaces/IBuilding";
 import ISavedBuilding from "../../interfaces/ISavedBuilding";
 
-import Badge from "react-bootstrap/Badge";
 import Card from "react-bootstrap/Card";
 import ListGroup from "react-bootstrap/ListGroup";
 import Tab from "react-bootstrap/Tab";
 import Tabs from "react-bootstrap/Tabs";
-import Stack from "react-bootstrap/Stack";
 
 export interface AllBuildingCardProps {
   building: IBuilding;
@@ -54,19 +49,6 @@ const BuildingCard: React.FC<AllBuildingCardProps> = ({
 
   const { currentUser, accountType } = useAuth();
 
-  // All Buildings Page - save/saved button
-  const [, /* modalState */ setModalState] = useContext(ModalContext);
-  const handleShowLogin = () => setModalState(ModalState.LOGIN);
-
-  function handleToggleSaveBuilding() {
-    if (savedHomeData) {
-      deleteBuilding(currentUser?.uid, buildingID, buildingName, accountType);
-    } else {
-      saveBuilding(currentUser?.uid, buildingID, buildingName, accountType);
-    }
-    shouldScroll.current = false;
-  }
-
   function getSentenceItem(): JSX.Element | null {
     if (!listing || listing.listingStatus !== listingStatusEnum.ACTIVE) {
       return (
@@ -85,48 +67,6 @@ const BuildingCard: React.FC<AllBuildingCardProps> = ({
       return null;
     }
   }
-
-  const header = (
-    <Card.Header>
-      <Card.Title className="mt-2">
-        <div>
-          {buildingName}
-          {willShowAvailTable(listing) && (
-            <Badge pill bg="warning" text="dark" className="units-avail-badge">
-              Units available!
-            </Badge>
-          )}
-        </div>
-      </Card.Title>
-      <Card.Subtitle>{address.neighborhood}</Card.Subtitle>
-      <div className="mt-2">
-        {currentUser ? (
-          savedHomeData ? (
-            <Stack direction={"horizontal"} gap={2}>
-              <WebsiteButton building={building} />
-              <SaveButton
-                isSaved={true}
-                onClickCallback={handleToggleSaveBuilding}
-              />
-            </Stack>
-          ) : (
-            <Stack direction={"horizontal"} gap={2}>
-              <WebsiteButton building={building} />
-              <SaveButton
-                isSaved={false}
-                onClickCallback={handleToggleSaveBuilding}
-              />
-            </Stack>
-          )
-        ) : (
-          <Stack direction={"horizontal"} gap={2}>
-            <WebsiteButton building={building} />
-            <SaveButton isSaved={false} onClickCallback={handleShowLogin} />
-          </Stack>
-        )}
-      </div>
-    </Card.Header>
-  );
 
   const expiresOrAgeRestricted =
     isEnding || isAgeRestricted ? (
@@ -174,7 +114,11 @@ const BuildingCard: React.FC<AllBuildingCardProps> = ({
       style={{ cursor: "pointer" }}
       onClick={() => setSelectedBuildingId?.(building.buildingID)}
     >
-      {header}
+      <BuildingCardHeader
+        building={building}
+        savedHomeData={savedHomeData}
+        shouldScroll={shouldScroll}
+      />
       <ListGroup variant="flush" className="mb-2">
         {getSentenceItem()}
         <ListGroup.Item>
